+++ crates/pms-types-payload/tests/general.rs
@@ -28,6 +28,7 @@ fn mvp_encrypt_decrypt_confidential_type_only() {
         PlainPayload::Genesis => println!("Type = Genesis"),
         PlainPayload::Mint { .. } => println!("Type = Mint"),
         PlainPayload::TxUtxo(_) => println!("Type = TxUtxo"),
+        PlainPayload::TokenBurn { .. } => println!("Type = TokenBurn"),
         PlainPayload::Milestone { .. } => println!("Type = Milestone"),
         PlainPayload::Nft(_) => println!("Type = Nft"),
         PlainPayload::ConfigUpdate(_) => println!("Type = ConfigUpdate"),